[feat] 🧚🏻‍♀️예적금 상품 상세 조회 api #58
